0.1.0
adjusted project structure - now utility namespaces of the components of the neural network are within the same directories as their respective actual class components. Status widget is now connected fully to the train method of the neural network so live data will be displayed on the GUI.  and network train method now makes use of validation split if the user chooses for it to be used.